Cleaned up the Front End
- Removed Js-webshim (Not Used)
- Renamed index.php to index.html
- Removed rankingstable.php (Not Used)
- Removed regional1.php (Not Used)
- Moved tomato.php to the proper location
