--- html/events/velocity-vortex/rankingstable.php
@@ -1,4 +1,0 @@
-<?php
-	require '../../potato.php';
-	RankingsTable('Y201702255');
-?>
