+++ html/index.html
@@ -1,0 +1,88 @@
+<!DOCTYPE html>
+<html>
+	<head>
+		<title>The Orange Alliance</title>
+		<link href="https://fonts.googleapis.com/css?family=Open+Sans" rel="stylesheet">
+		<meta name = "viewport" content="width=device-width, initial-scale=1.0">
+		<link href = "css/bootstrap.min.css" rel = "stylesheet" type="text/css">
+		<link href = "css/styles.css" rel = "stylesheet" type="text/css">
+		<link href = "css/jquery.dataTables.min.css" rel = "stylesheet" type="text/css">
+		<link rel="apple-touch-icon" sizes="152x152" href="https://sites.google.com/site/filehostdummysite1234/files/theorangealliance/apple-touch-icon.png?v=vMrqOno5qk">
+		<link rel="icon" type="image/png" href="https://sites.google.com/site/filehostdummysite1234/files/theorangealliance/favicon-32x32.png?v=vMrqOno5qk" sizes="32x32">
+		<link rel="icon" type="image/png" href="https://sites.google.com/site/filehostdummysite1234/files/theorangealliance/favicon-16x16.png?v=vMrqOno5qk" sizes="16x16">
+		<link rel="manifest" href="https://sites.google.com/site/filehostdummysite1234/files/theorangealliance/manifest.json?v=vMrqOno5qk">
+		<link rel="mask-icon" href="https://sites.google.com/site/filehostdummysite1234/files/theorangealliance/safari-pinned-tab.svg?v=vMrqOno5qk" color="#ff9500">
+		<link rel="shortcut icon" href="https://sites.google.com/site/filehostdummysite1234/files/theorangealliance/favicon.ico?v=vMrqOno5qk">
+		<meta name="apple-mobile-web-app-title" content="The Orange Alliance">
+		<meta name="application-name" content="The Orange Alliance">
+		<meta name="msapplication-config" content="https://sites.google.com/site/filehostdummysite1234/files/theorangealliance/browserconfig.xml?v=vMrqOno5qk">
+		<meta name="theme-color" content="#ff9500">
+		<meta http-equiv="Content-Type" content="text/html;charset=utf-8" />
+		<script src = "http://ajax.googleapis.com/ajax/libs/jquery/1.10.2/jquery.min.js"></script>
+		<script src="js/jquery.dataTables.min.js"></script>
+		<script src="js/dataTables.bootstrap.min.js"></script>
+		<script type="text/javascript">
+		$(document).ready(function() { 
+		$("#inputTable1").DataTable({
+			paging: false
+			}); 
+		$("#inputTable2").DataTable({
+			paging: false
+			}); 
+		$("#inputTable3").DataTable({
+			paging: false
+			}); 
+			});
+		</script>
+	</head>	
+	<body style= "padding-top: 0px;">
+		<div style = "margin-bottom: 0px;" class="navbar navbar-default navbar-fixed-top">
+			<div class="container">
+				<div class="navbar-header">
+					<a class="nav-brand" href="http://theorangealliance.tk:8080/"> 
+					<img style="max-width:50px" src="/images/logo.png"> 
+					<span class="logo hidden-xs">The Orange Alliance</span>
+					</a>
+					<a class="nav-brand" href="http://theorangealliance.tk:8080/"></a>
+					<button class = "navbar-toggle" data-toggle = "collapse" data-target = ".navHeaderCollapse">
+						<span class = "icon-bar"></span>
+						<span class = "icon-bar"></span>
+						<span class = "icon-bar"></span>
+					</button>
+				</div>
+				<div class = "collapse navbar-collapse navHeaderCollapse">
+					<ul class = "nav navbar-nav navbar-right">
+						<li class = "active"><a href = "http://theorangealliance.tk:8080/">Home</a></li>
+						<li ><a href = "http://theorangealliance.tk:8080/events/velocity-vortex.php">Events</a></li>
+						<li ><a href = "http://theorangealliance.tk:8080/input-data.php">Input Match Results</a></li>
+					</ul>
+				</div>
+			</div>
+		</div>
+		<div class="content">
+				<div class="header">
+				<center><img class="img-responsive" src="/images/logotext.png" style="padding-top: 200px"></center>
+				</div>
+		</div>
+		<div class="footer">
+			<div class="container">
+				<div class="col-md-6" style="padding-bottom: 10px;">
+				<center>
+				Designed by:
+				Team 8097 Botcats,
+				Team 9261 Level Up,
+				Team 10809 Crow Force 5
+				</center>
+				</div>
+				<div class="col-md-3">
+				</div>
+				<div class="col-md-3" >
+				<center>
+					© TheOrangeAlliance 2017
+					</cetner>
+				</div>
+			</div>
+		</div>
+		<script src = "js/bootstrap.js"></script>
+	</body>
+</html>
